Refactor: 그룹 통계 관련 클래스 유틸 클래스로 전환
Ref: #79
