+++ src/main/java/com/studypals/domain/groupManage/util/GroupStudyStatisticCalculator.java
@@ -1,6 +1,9 @@
-package com.studypals.domain.groupManage.worker;
+package com.studypals.domain.groupManage.util;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -8,16 +11,14 @@ import com.studypals.domain.groupManage.dto.*;
 import com.studypals.domain.groupManage.entity.Group;
 import com.studypals.domain.groupManage.entity.GroupStudyCategory;
 import com.studypals.domain.studyManage.dto.GetStudyOfMemberDto;
-import com.studypals.global.annotations.Worker;
 
 /**
- * 그룹의 공부 시간에 대한 통계 관련 로직을 담당하는 Worker 입니다.
+ * 그룹의 공부 시간을 조합하고 통계 결과를 계산하는 유틸성 클래스입니다.
  *
  * @author s0o0bn
- * @since 2025-05-10
+ * @since 2025-05-21
  */
-@Worker
-public class GroupStudyStatisticManager {
+public class GroupStudyStatisticCalculator {
     private static final int SECONDS_PER_MINUTE = 60;
 
     /**
@@ -27,7 +28,7 @@ public class GroupStudyStatisticManager {
      * @param studies 그룹원의 공부 정보에 대한 DTO {@link GetStudyOfMemberDto}
      * @return 그룹 카테고리마다 그룹원 별 공부 시간 총합 DTO {@link GroupTotalStudyDto}
      */
-    public GroupTotalStudyDto sumTotalTimeOfCategory(
+    public static GroupTotalStudyDto sumTotalTimeOfCategory(
             List<GroupMemberProfileDto> members, List<GetStudyOfMemberDto> studies) {
         Map<Long, GroupMemberProfileDto> memberMap =
                 members.stream().collect(Collectors.toMap(GroupMemberProfileDto::id, Function.identity()));
@@ -56,7 +57,7 @@ public class GroupStudyStatisticManager {
      * @param categories 그룹의 공부 카테고리 목록
      * @return 목표 달성률, 달성한 그룹원 프로필을 포함한 카테고리 정보 DTO 목록 {@link DailySuccessRateRes}
      */
-    public List<DailySuccessRateDto> getDailySuccessRate(
+    public static List<DailySuccessRateDto> getDailySuccessRate(
             Group group, GroupTotalStudyDto totalStudy, List<GroupStudyCategory> categories) {
         return categories.stream()
                 .map(category -> getCategorySuccessRate(category, totalStudy, group.getTotalMember()))
@@ -71,7 +72,7 @@ public class GroupStudyStatisticManager {
      * @param totalMember 전체 그룹원 수
      * @return 해당 카테고리의 {@link DailySuccessRateRes}
      */
-    private DailySuccessRateDto getCategorySuccessRate(
+    private static DailySuccessRateDto getCategorySuccessRate(
             GroupStudyCategory category, GroupTotalStudyDto totalStudy, int totalMember) {
         Map<GroupMemberProfileDto, Long> memberStudiedTime =
                 totalStudy.memberTotalStudiedTimePerCategory().getOrDefault(category.getId(), Collections.emptyMap());
